+++ docs/Whats-New.md
@@ -542,6 +542,7 @@ New:
 - Spawns particle when spawns tiberium by terrain (by NetsuNegi)
 - Allow jumpjet climbing ignore building height (by TaranDahl)
 - Allow draw SuperWeapon timer as percentage (by NetsuNegi)
+- Customize particle system of parasite logic (by NetsuNegi)
 - DeployFire supports buildings (By FlyStar)
 
 Vanilla fixes:
@@ -601,6 +602,7 @@ Vanilla fixes:
 - Fixed the bug that cause technos teleport to cell 0,0 by ChronoSphere superweapon (by NetsuNegi)
 - Fixed the bug that techno in attack move will move to target if it cannot attack it (by NetsuNegi)
 - Fixed the bug in AI scripts 56 and 57 that forced the launch of superweapons with index numbers 3 and 4 (by FlyStar)
+- Buildings with `NeedsEngineer=true` are now considered to have threat value of 0 under ownership of `MultiplayPassive=true` houses regardless of their `ThreatPosed` value (by Starkku)
 
 Phobos fixes:
 - Fixed the bug that `AllowAirstrike=no` cannot completely prevent air strikes from being launched against it (by NetsuNegi)